rename minamed exception class and create package
