--- src/main/java/com/cqse/teamscaleupload/AgentOptionParseException.java
@@ -1,27 +1,0 @@
-package com.cqse.teamscaleupload;
-
-/**
- * Thrown if option parsing fails.
- */
-public class AgentOptionParseException extends Exception {
-
-    /**
-     * Serialization ID.
-     */
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * Constructor.
-     */
-    public AgentOptionParseException(String message) {
-        super(message);
-    }
-
-    /**
-     * Constructor.
-     */
-    public AgentOptionParseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
